changed weight type in schema, had to drop table and add it back because its sqlite

+++ pups3/models.py
@@ -1,5 +1,5 @@
-from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy import Column, Integer, String, ForeignKey, Date, Numeric
+from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 from sqlalchemy import create_engine
 
@@ -24,7 +24,7 @@ class Puppy(Base):
     picture = Column(String)
     shelter_id = Column(Integer, ForeignKey('shelter.id'))
     shelter = relationship(Shelter)
-    weight = Column(Numeric(10))
+    weight = Column(Integer)
 
 engine = create_engine("sqlite:///puppyshelter.db")
